pyproject.toml
[project]
name = "minerbot"
version = "0.1.0"
description = "AI Assistant built with LangChain DeepAgents - CLI Personal AI Assistant"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "pyyaml>=6.0",
    "deepagents>=0.4.7",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0",
    "typer>=0.12.0",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "aiosqlite>=0.22.1",
]

[project.scripts]
minerbot = "minerbot.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.408",
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]
[tool.hatch.build.targets.wheel]
packages = ["src/minerbot"]
